+++ AirHabitat.java
@@ -1,7 +1,14 @@
 // File : AirHabitat.java
 // Author : Prama Legawa Halqavi (13515132)
-
+ /**
+  * Kelas yang menunjukkan habitat udara
+  * @author Prama Legawa Halqavi/13515132
+  * @version 1.0, March 2017
+  */
 public class AirHabitat extends Habitat {
+  /**Constructor AirHabitat
+    *@see java.lang.reflect.Constructor
+	  */
   public AirHabitat() {
     super("Air");
   }
