+++ AnimalClass.java
@@ -1,6 +1,17 @@
+
 // Nama :  M. Ferdi Ghozali
 // NIM : 13515014
+/** Class yang menunjukkan kelas hewan
+  * @author FerdiGhozali/13515014
+  * @version 1.0, March 2017
+  */
 interface AnimalClass {
+  /** Menghasilkan deskripsi tentang animal
+    * 
+    */
   public String GetDeskripsi();
+  /** Menghasilkan nama kelas animal
+    * 
+    */
   public String GetClassName();
 }
