+++ AnimalFoodType.java
@@ -1,5 +1,13 @@
+
 // Nama :  M. Ferdi Ghozali
 // NIM : 13515014
+/** Class yang menunjukkan type makanan hewan
+  * @author FerdiGhozali/13515014
+  * @version 1.0, March 2017
+  */
 interface AnimalFoodType {
+  /** Menghasilkan type makanan hewan 
+    * 
+    */
   public String GetFoodType();
 }
