+++ Cage.java
@@ -1,8 +1,12 @@
 //File: Cage.java
 //Author: David Theosaksomo
 //NIM: 13515131
-
 import java.util.Random;
+/**
+ * Class yang menunjukkan Cage atau kandang
+ * @author David Theosaksomo/13515131
+ * @version 1.0, March 2017
+ */
 public class Cage implements Renderable {
 
   protected final int id;
@@ -12,10 +16,18 @@ public class Cage implements Renderable {
   protected int jumlah_hewan = 0;
   protected final int animal_size = 50;
   protected Animal[] list_of_animal;
-
+  /** Constructor
+   *  @see java.lang.reflect.Constructor
+   */
   public Cage() {
     id = 0;
   }
+  /** Constructor dengan parameter
+   *  @param id sebagai penanda id Cage
+   *  @param area sebagai penanda point yang ada pada Cage
+   *  @param ukuran sebagai penanda ukuran Cage
+   *  @param type sebagai penanda type Cage
+   */
   public Cage(int id, Point[] area, int ukuran, String type) {
     this.id = id;
     cage_type = type;
@@ -26,24 +38,42 @@ public class Cage implements Renderable {
     }
     list_of_animal = new Animal[animal_size];
   }
+  /**  Menghasilkan nilai dari ukuran Cage
+   */
   public int GetCageSize() {
     return size_area;
   }
+  /** Menghasilkan ID Cage
+  */
   public int GetID() {
     return id;
   }
+  /** Menghasilkan jumlah hewan pada Cage
+   */
   public int NumberOfAnimal() {
     return jumlah_hewan;
   }
+  /** Menghasilkan type Cage
+   */
   public String GetTypeCage() {
     return cage_type;
   }
+  /** Menghasilkan Point ke i di dalam Cage
+   * @param i merupakan indeks Point yang akan dicari
+   */
   public Point GetAreaElmt(int i) {
     return area[i];
   }
+  /** Menghasilkan Animal ke i pada list_of_animal
+   * @param i merupakan indeks list_of_animal yang akan dicari
+   */
   public Animal GetAnimal(int i ) {
     return list_of_animal[i];
   }
+  /** Memasukkan Animal ke dalam Cage
+   * @param hewan menunjukkan hewan yang akan dimasukkan
+   * @param area_id ID dari area yang akan dimasukkan hewan
+   */
   public void InputAnimal(Animal hewan, int area_id) {
     if (Aman(hewan) && (0.3*size_area - (jumlah_hewan+1) > 0) && 
        (hewan.IsTypeThere(cage_type))) {
@@ -54,6 +84,9 @@ public class Cage implements Renderable {
         }
     }
   }
+  /** Mengecek hewan apakah aman bila dimasukkan ke Cage
+   * @param hewan menunjukkan hewan yang akan dimasukkan
+   */
   public boolean Aman(Animal hewan) {
     boolean safe = true;
     if (!hewan.IsTamed()){
@@ -78,6 +111,9 @@ public class Cage implements Renderable {
     }
     return(safe);
   }
+  /** Mengecek apakah dapat melewati point
+   * @param p merupakan Point yang akan dicek
+   */
   public boolean IsPassAble(Point p) {
     boolean bener = false;
     for (int i = 0; i < size_area;i++) {
@@ -95,6 +131,9 @@ public class Cage implements Renderable {
     }
     return bener;    
   }
+  /** Menggerakkan Animal
+   * @param p merupakan Point yang akan dicek
+   */
   public void MoveAnimal() {
     Random rand = new Random();
    
@@ -123,6 +162,9 @@ public class Cage implements Renderable {
       }
     }
   }
+  /** Mencetak ID dari Cage
+   * @see Renderable#Render()
+   */
   public void Render() {
     System.out.print(id);
   }
