+++ Driver.java
@@ -1,12 +1,15 @@
 //File: Driver.java
 //Author: David Theosaksomo
 //NIM: 13515131
-
-import java.util.Scanner; 
+/**
+  * Class yang merupakan driver program
+  * @author David Theosaksomo/13515131
+  * @version 1.0, March 2017
+  */
+import java.util.Scanner;
 import java.util.Random;
 
-/** @class Zoo
-  * Kelas zoo memiliki atribut  matriks Cell dimana 1 cell merepresentasikan 
+/** Kelas Zoo memiliki atribut  matriks Cell dimana 1 cell merepresentasikan 
   * sebuah petak 1x1m
   */
 public class Driver {
@@ -24,7 +27,9 @@ public class Driver {
   private Point guest_pos;
   private int n_animal = 0;
   private boolean[] is_ada_cage;
-
+  /** Driver program Virtual Zoo
+   * 
+   */
   public Driver(){
     virtual_zoo = new Zoo(default_width,default_height);
     guest_pos = new Point(-1,-1);
@@ -263,15 +268,18 @@ public class Driver {
 
 
   }
-   /** @brief Menampilkan Zoo diatas layar dengan method render.
-     * @param upleft kordinat atas kiri
-     * @param downright kordinat bawah kanan
-     */
+  
+  /** Menampilkan Zoo diatas layar dengan method render.
+   */
   public void DisplayVirtualZoo(){
     Point up = new Point(0,0);
     Point down = new Point(width-1,height-1);
     DisplayVirtualZoo(up,down);
   }
+  /** Menampilkan Zoo diatas layar dengan method render.
+   * @param upleft kordinat atas kiri
+   * @param downright kordinat bawah kanan
+   */
   public void DisplayVirtualZoo(Point upleft, Point downright){
      //Inisialisasi variabel antara
     int[][] cage_matrix = new int[height][width];
@@ -355,6 +363,8 @@ public class Driver {
         System.out.println('|');
     }
   }
+  /** Menghitung jumlah makanan seluruh hewan
+   */
   public void HitungMakanan(){
     double daging = 0;
     double sayur = 0;
@@ -385,6 +395,8 @@ public class Driver {
     System.out.println(sayur + " kg sayur, dan");
     System.out.println(omni + " kg daging/sayur,");  
   }
+  /** Menggerakan hewan
+   */
   public void MoveAnimal(){
     for (int i = 0; i < max_cage; i++) {
       if (is_ada_cage[i]) {
@@ -392,6 +404,8 @@ public class Driver {
       }
     }
   }
+  /** Mencetak Judul Program Virtual Zoo
+   */
   public void PrintTitle(){
     System.out.println(" _____     _ _    _               ______        ");    
     System.out.println("|_   _|   | | |  (_)             |___  /            ");
@@ -403,7 +417,7 @@ public class Driver {
     System.out.println("                          |___/   ");
     System.out.println("========================================================");  
   }
-   /** @brief  secara acak memilih salah satu Entrance, kemudian membuat sebuah 
+   /** Secara acak memilih salah satu Entrance, kemudian membuat sebuah 
      * jalur tour yang menampilkan serangkaian
      * experience yang akan dialami pengunjung  berdasarkan interaksi dengan
      * hewan-hewan yang dilalui tour tersebut. 
@@ -649,7 +663,10 @@ public class Driver {
       }
     }
     System.out.println("\nTour selesai.");
+    scanner.close();
   }
+  /** Mencetak seluruh hewan dalam map
+   */
   public void PrintAllAnimal(){
     int count = 0;
     for (int i = 0; i < max_cage; i++) {
