+++ Entrance.java
@@ -1,5 +1,12 @@
+/**Class Entrance menunjukkan pintu masuk
+ * @author Prama Legawa Halqavi/13515132
+ * @version 1.0/March 2017
+ */
 public class Entrance extends Road {
-  // ctor tanpa parameter
+  /**Constructor
+   * Memanggil method constructor parents
+   * dengan parameter "Entrance"
+   */
   public Entrance() {
     super("Entrance");
   }
