+++ Exit.java
@@ -1,5 +1,12 @@
+/**Class Entrance menunjukkan pintu keluar
+ * @author Prama Legawa Halqavi
+ * @version 1.0/March 2017
+ */
 public class Exit extends Road {
-  // ctor tanpa parameter
+  /**Constructor
+   * Memanggil method constructor parents
+   * dengan parameter "Exit"
+   */
   public Exit() {
     super("Exit");
   }
