+++ Facility.java
@@ -1,3 +1,8 @@
+/**
+  * Class yang menunjukkann bawain dari iif
+  * @author Prama Legawa Halqavi/13515132
+  * @version 1.0, March 2017
+  */
 public abstract class Facility extends Cell {
   protected final String facility_name; //  nama fasilitas
   protected final String facility_type; //  nama fasilitas
@@ -5,36 +10,41 @@ public abstract class Facility extends Cell {
   protected static String defname = "The Facility";
   protected static int default_max_pengunjung = 10;
   protected static String nil_string = "Nil";
-  /* * @brief ctor tanpa parameter. 
-       * nama dan tipe fasilitas diinisialisasi dengan defname */
+  /**ctor tanpa parameter, 
+   *nama dan tipe fasilitas diinisialisasi dengan defname
+   */
   public Facility() {
     super("Facility");
     this.facility_name = defname;
     this.facility_type = defname;
   }
-  /* * @brief ctor dengan parameter 
-     * @param name nama fasilitas 
-     * @param type tipe fasilitas */
+  /** ctor dengan parameter 
+   * @param name nama fasilitas 
+   * @param type tipe fasilitas 
+   */
   public Facility(String name, String type) {
     super("Facility");
     this.facility_name = name;
     this.facility_type = type;
     cell_type = "Facility";
   }
-  /* * @brief mengakses type fasilitas 
-     */
+  /**mengakses type fasilitas 
+    */
   public final String GetFacHabType() {
     return facility_type;
   }
- /* * @brief mengakses nama fasilitas */
+ /**mengakses nama fasilitas 
+   */
   public final String GetFacilityName() {
     return facility_name;
   }
- /* * @brief menggambarkan objek ke layar */
+ /**menggambarkan objek ke layar 
+   */
   public void Render(boolean oncage) {
     Render();
   }
-  /* * @brief menggambarkan objek ke layar */
+  /**menggambarkan objek ke layar 
+    */
   public void Render() {
     if (facility_type.equals("Park")) {
       System.out.print("P");
