+++ Habitat.java
@@ -1,13 +1,30 @@
+//Nama : Radiyya Dwisaputra
+//NIM : 13515023
+/**
+  * Class yang menunjukkan habitat
+  * @author Radiyya Dwisaputra/13515023
+  * @version 1.0, March 2017
+  */
 public class Habitat extends Cell {
 	protected final String habitat_type;
-
+	/** Constructor
+   *  @see java.lang.reflect.Constructor
+   */
 	public Habitat(String s) {
 		super("Habitat");		
 		habitat_type=s;
 	}
+	/**
+	 * Menghasilkan type habitat
+	 * @see Cell#GetFacHabType()
+	 */
 	public String GetFacHabType() {
 		return habitat_type;
 	}
+	/**
+	 * Mencetak symbol di map
+	 * @see Cell#Render()
+	 */
 	public void Render() {
 		if(habitat_type == "Water") {
 			System.out.print ("w");
@@ -19,6 +36,11 @@ public class Habitat extends Cell {
 			System.out.print ("a");
 		}
 	}
+	/**
+	 * Mencetak habitat pada Cage
+	 * @param oncage ialah habitat terletak dalam Cage atau tidak
+	 * @see Cell#Render(boolean)
+	 */
 	public void Render(boolean oncage) {
 		if(oncage) {
 			if(habitat_type == "Water") {
@@ -34,6 +56,10 @@ public class Habitat extends Cell {
 		}
 		}
 	}
+	/**
+   * Meletakkan pintu masuk di map
+	 * @see Cell#Enter()
+	 */
 	public void Enter() {
 	}
 }
