+++ HiuGoblin.java
@@ -1,5 +1,12 @@
+
+
 //Nama : Radiyya Dwisaputra
 //NIM : 13515023
+/**
+  * Class yang menunjukkan hewan real HiuGoblin
+  * @author Radiyya Dwisaputra/13515023
+  * @version 1.0, March 2017
+  */
 public class HiuGoblin extends Animal
 implements AnimalClass, AnimalFoodType
 {
@@ -7,6 +14,9 @@ implements AnimalClass, AnimalFoodType
   private final String food_type = "Carnivore";
   private final String kelas = "Chondrichtyes";
   private final String deskripsi = "Chondrichthyes merupakan ikan dengan tulang rawan";
+  /** Constructor
+   *  @see java.lang.reflect.Constructor
+   */
   public HiuGoblin() {
     super('h',false);
     weight = defaultweight;
@@ -16,22 +26,40 @@ implements AnimalClass, AnimalFoodType
     species_name = "HiuGoblin";
     animal_type[0] = "Water";
   }
+  /** Interaction HiuGoblin kepada pengunjung
+   * @see Animal#Interact()
+   */
    public void Interact() {
     System.out.println( "I can extend my jaws!");
   }
+  /** Pengecekan apakah String animal enemy atau bukan
+   * @param animal string animal yang akan dicek
+   * @see Animal#IsEnemy(java.lang.String)
+   */
   public boolean IsEnemy(String animal) {
     return false;
   }
+  /** Memperoleh type makanan animal
+   * @see Animal#GetFoodType()
+   * @see AnimalFoodType#GetFoodType()
+   */
   public String GetFoodType(){
     return food_type;
   }
+  /** Memperoleh nama kelas dari animal
+   * @see AnimalClass#GetClassName()
+   */
   public String GetClassName(){
     return kelas;
   }
+  /** Memperoleh deskripsi animal 
+   * @see AnimalClass#GetDeskripsi()
+   */
   public String GetDeskripsi(){
     return deskripsi;
   }
-
+  /** Menuliskan deskripsi hewan
+   */
   public void Deskripsi(){
     System.out.print(super.GetSpeciesName());
     System.out.print(" termasuk dalam kelas ");
