+++ Kungkang.java
@@ -1,7 +1,8 @@
 //Nama : Radiyya Dwisaputra
 //NIM : 13515023
-/** @brief Kelas objek Kungkang yang mewakili hewan Kungkang. 
-  * objek akan di masukkan ke dalam zoo.
+/** Class yang menunjukkan hewan real Kungkang
+  * @author Radiyya Dwisaputra/13515023
+  * @version 1.0, March 2017
   */
 public class Kungkang extends Animal
 implements AnimalClass, AnimalFoodType
@@ -11,9 +12,9 @@ implements AnimalClass, AnimalFoodType
   private final String kelas = "Mammal";
   private final String deskripsi = "Mammal merupakan kelas hewan vertebrata" + 
   "yang ciri utama nya memiliki kelenjar susu";
-  /** @brief Ctor objek Kungkang.
-    * Terbentuk objek Kungkang.
-    */ 
+  /** Constructor
+   *  @see java.lang.reflect.Constructor
+   */
   public Kungkang() {
     super('k',true);
     weight = defaultweight;
@@ -23,42 +24,46 @@ implements AnimalClass, AnimalFoodType
     species_name = "Kungkang";
     animal_type[0] = "Land";
   } 
-  /** @brief Method Interact objek Kungkang.
-    * Menampilkan keluaran saat adanya interaksi.
-    * Interaksi dilakukan dengan pengunjung.
-    */
+  /** Interaction Kungkang kepada pengunjung
+   * @see Animal#Interact()
+   */
    public void Interact() {
     System.out.println("aku gak gendut, cuman agak bulet dikit");
     System.out.println("Mammal merupakan kelas hewan vertebrata" + 
     "yang ciri utama nya memiliki kelenjar susu");
   }
-  /** @brief Method IsEnemy objek Kungkang.
-    * Mengecek hewan lain apakah musuh Kungkang atau tidak.
-    * Musuh ialah yang tidak dapat diletakkan 1 kandang.
-    * @param hewan yang akan di cek.
+   /** Pengecekan apakah String animal enemy atau bukan
+    * @param animal string animal yang akan dicek
+    * @see Animal#IsEnemy(java.lang.String)
     */
-  public boolean IsEnemy(String animal) {
-    return false;
-  }
-  /** @brief Method GedFoodType objek Kungkang.
-    * Memperoleh tipe makanan yang dimakan oleh hewan.
+   public boolean IsEnemy(String animal) {
+     return false;
+   }
+   /** Memperoleh type makanan animal
+    * @see Animal#GetFoodType()
+    * @see AnimalFoodType#GetFoodType()
     */
-  public String GetFoodType(){
-    return food_type;
-  }
-  public String GetClassName(){
-    return kelas;
-  }
-  public String GetDeskripsi(){
-    return deskripsi;
-  }
-  /** @brief Method Deskripsi objek Kungkang.
-      * Menghasilkan deskripsi tentang hewan.
-      */
-  public void Deskripsi(){
-    System.out.print(super.GetSpeciesName());
-    System.out.print(" termasuk dalam kelas ");
-    System.out.println(GetClassName());
-    System.out.println(deskripsi);
-  }
+   public String GetFoodType(){
+     return food_type;
+   }
+   /** Memperoleh nama kelas dari animal
+    * @see AnimalClass#GetClassName()
+    */
+   public String GetClassName(){
+     return kelas;
+   }
+   /** Memperoleh deskripsi animal 
+    * @see AnimalClass#GetDeskripsi()
+    */
+   public String GetDeskripsi(){
+     return deskripsi;
+   }
+   /** Menuliskan deskripsi hewan
+    */
+   public void Deskripsi(){
+     System.out.print(super.GetSpeciesName());
+     System.out.print(" termasuk dalam kelas ");
+     System.out.println(GetClassName());
+     System.out.println(deskripsi);
+   }
 }
