+++ LandHabitat.java
@@ -1,7 +1,13 @@
 // File : LandHabitat.java
 // Author : Prama Legawa Halqavi (13515132)
-
+/** Class yang menunjukkan land habitat
+  * @author Prama Legawa Halqavi/13515132
+  * @version 1.0, March 2017
+  */
 public class LandHabitat extends Habitat {
+  /** Constructor
+   * @see java.lang.Constructor
+   */
   public LandHabitat() {
     super("Land");
   }
