+++ Okapi.java
@@ -1,3 +1,5 @@
+
+
 //Nama : Radiyya Dwisaputra
 //NIM : 13515023
 /** @brief Kelas objek Okapi yang mewakili hewan Okapi.
