+++ Platypus.java
@@ -1,3 +1,5 @@
+
+
 //Nama : Radiyya Dwisaputra
 //NIM : 13515023
 /** @brief Kelas objek Platypus yang mewakili hewan Platypus.
