+++ Renderable.java
@@ -1,6 +1,4 @@
-//File: Renderable.java
-//Author: David Theosaksomo
-//NIM: 13515131
+
 
 interface Renderable{
   public void Render();
