+++ SemutPanda.java
@@ -1,3 +1,5 @@
+
+
 //Nama : Radiyya Dwisaputra
 //NIM : 13515023
 /** @brief Kelas objek SemutPanda yang mewakili hewan SemutPanda.
